Move training logic into base class
